requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0

# LLM Providers
openai>=1.50.0
google-generativeai>=0.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# UI
streamlit>=1.38.0

# Environment and Config
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
tenacity>=8.0.0
